requirements.txt
playwright==1.32.1
pytest==7.2.0
python-dotenv==1.0.0
pytest-playwright==0.3.2
pytest-xdist==3.2.1
allure-pytest==2.13.2
pytest-html==3.2.0
flaky==3.7.0